+++ nes-runtime/src/Execution/Operators/Streaming/EventTimeWatermarkAssignment.cpp
@@ -24,7 +24,7 @@ namespace NES::Runtime::Execution::Operators {
 class WatermarkState : public OperatorState {
   public:
     explicit WatermarkState() {}
-    Value<> currentWatermark = Value<UInt64>((uint64_t) 0);
+    Value<> currentWatermark = Value<UInt64>(0_u64);
 };
 
 EventTimeWatermarkAssignment::EventTimeWatermarkAssignment(TimeFunctionPtr timeFunction)
